requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
openai==1.5.0
langchain==0.0.350
tabulate==0.8.10